test(mcp): move config completeness spec to canonical location and refactor
